% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers_clinical.R
\name{classify_g6pd}
\alias{classify_g6pd}
\title{Classify G6PD enzyme activity}
\usage{
classify_g6pd(g6pd)
}
\arguments{
\item{g6pd}{Numeric vector of G6PD activity in U/gHb.
Values ≥ 6.1 are classified as normal.}
}
\value{
A character vector: \code{"normal G6PD"} or \code{"abnormal G6PD"}.
}
\description{
Classify G6PD enzyme activity
}
\examples{
classify_g6pd(c(7.0, 4.5, NA))

}
