% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers_demographic.R
\name{classify_job_type}
\alias{classify_job_type}
\title{Collapse raw job categories into three employment groups}
\usage{
classify_job_type(x)
}
\arguments{
\item{x}{Character or factor vector of raw job categories.}
}
\value{
A factor with four levels:
\code{"Formal Employment"}, \code{"Informal Employment"},
\code{"House wife"}, \code{"Not working"}.
}
\description{
Collapses the granular \code{job_scorres} field into Formal Employment,
Informal Employment, Housewife, or Not Working.
}
\examples{
classify_job_type(c("Salaried worker", "Small business", "Housewife",
                    "Not applicable/Not working"))

}
