% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers_demographic.R
\name{classify_maternal_age}
\alias{classify_maternal_age}
\title{Classify maternal age into clinical categories}
\usage{
classify_maternal_age(age)
}
\arguments{
\item{age}{Numeric vector of maternal ages in years.}
}
\value{
An ordered factor with levels:
\itemize{
\item \code{"Teenage (<20)"}
\item \code{"Adult (20-<35)"}
\item \code{"Advanced Maternal Age (35+)"}
}
}
\description{
Classify maternal age into clinical categories
}
\examples{
classify_maternal_age(c(17, 24, 36, NA))

}
