% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers_demographic.R
\name{classify_residence}
\alias{classify_residence}
\title{Classify facility location as rural or peri-urban}
\usage{
classify_residence(facility_name)
}
\arguments{
\item{facility_name}{Character vector of facility name strings
(e.g. \code{scrn_fac_spfy_obsloc}).}
}
\value{
A factor with reference level \code{"Peri-urban Manyata"}.
}
\description{
Uses a pattern match on the facility name to assign a residence category.
Currently recognises Siaya / Tingwangi as Rural-Karemo and all other sites
as Peri-urban Manyata.
}
\examples{
classify_residence(c("Siaya County Referral Hospital",
                     "Kisumu County Hospital",
                     "Tingwangi Health Centre"))

}
