% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers_clinical.R
\name{classify_stfr}
\alias{classify_stfr}
\title{Classify iron status using soluble transferrin receptor (sTfR)}
\usage{
classify_stfr(stfr, trimester)
}
\arguments{
\item{stfr}{Numeric vector of sTfR in mg/L.}

\item{trimester}{Character vector of trimester labels (see \code{\link[=classify_trimester]{classify_trimester()}}).}
}
\value{
An ordered factor: \code{"none deficient"} or \code{"iron deficient"}.
}
\description{
Applies trimester-specific sTfR cut-offs to identify iron deficiency.
}
\details{
\tabular{ll}{
   Trimester \tab sTfR threshold for iron deficiency \cr
   1st \tab > 3.61 mg/L \cr
   2nd \tab > 4.98 mg/L \cr
   3rd \tab > 5.94 mg/L \cr
}
}
\examples{
classify_stfr(c(3.0, 5.5, 6.5),
              c("1st Trimester", "2nd Trimester", "3rd Trimester"))

}
