% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifiers_clinical.R
\name{derive_any_anemia}
\alias{derive_any_anemia}
\title{Derive binary any-anemia indicator (trimester-adjusted)}
\usage{
derive_any_anemia(hb, trimester)
}
\arguments{
\item{hb}{Numeric vector of haemoglobin concentration in g/dL.}

\item{trimester}{Character vector of trimester labels as returned by
\code{\link[=classify_trimester]{classify_trimester()}}, i.e. one of \code{"1st Trimester"},
\code{"2nd Trimester"}, \code{"3rd Trimester"}.}
}
\value{
An integer vector of 0/1.
}
\description{
Returns \code{1L} if Hb is below the trimester-specific threshold, \code{0L}
otherwise. Thresholds: < 10.5 g/dL in 2nd trimester, < 11.0 g/dL
in 1st and 3rd trimesters.
}
\examples{
derive_any_anemia(c(9.0, 11.5, 10.3), c("2nd Trimester",
  "1st Trimester", "3rd Trimester"))

}
